Add files via upload
Uploaded to JLC.   Branch ready to merge back to main.
Move edge to the left and change width to 99.5mm.   After checking a 100mm board in a 1455 case, I made this board .5mm wider.   Move the edges to give the transformer a little more clearance. 
Moved the caps at the bottom of the board up 1mm.  That is to give more clearance for output connectors.    5-way binding posts should clear, but this gives a little more leeway.